+++ src/js/components/common/VideosCardContainer.js
@@ -1,10 +1,10 @@
-import { DELETE_CONFIRM_MESSAGE } from '../constants';
+import { DELETE_CONFIRM_MESSAGE } from '../../constants';
 import {
   getStorageVideos,
   checkVideoStorageFull,
   setStorageVideos,
   removeStorageVideo
-} from '../utils/localStorage';
+} from '../../utils/localStorage';
 import TEMPLATE from './template';
 import toast from './toast';
 
@@ -91,7 +91,7 @@ export default class VideoCardContainer {
   };
 
   removeHandler = (e) => {
-    if (e.target.className.includes('video-item__remove-button') && confirm(DELETE_CONFIRM_MESSAGE)) {
+    if (e.target.className.includes('video-item__remove-button') && window.confirm(DELETE_CONFIRM_MESSAGE)) {
       const li = e.target.closest('li');
       const { videoId } = li.dataset;
       const { filter } = this.#state;
