+++ src/js/components/common/template.js
